Move publish and unpublish callback before s3 operations, so that we don't need to rollback s3 changes when these callbacks failed. And change the passed in db with tx, we can rollback all db changes if s3 failed.

+++ unpublish.go
@@ -25,12 +25,15 @@ func Unpublish(db *gorm.DB, version QorMicroSiteInterface, printActivityLog bool
 			return
 		}
 
+		if err1 = version.UnPublishCallBack(tx, version.GetMicroSiteURL()); err1 != nil {
+			return
+		}
+
 		for _, o := range version.GetFilesPathWithSiteURL() {
 			oss.Storage.Delete(o)
 		}
 
-		return version.UnPublishCallBack(db, version.GetMicroSiteURL())
-
+		return
 	})
 
 	return
